+++ src/frequenz/quantities/__init__.py
@@ -24,17 +24,17 @@ the quantity.
 
 This library provides the following types:
 
+- [ApparentPower][frequenz.quantities.ApparentPower]: A quantity representing apparent
+  power.
 - [Current][frequenz.quantities.Current]: A quantity representing an electric current.
 - [Energy][frequenz.quantities.Energy]: A quantity representing energy.
 - [Frequency][frequenz.quantities.Frequency]: A quantity representing frequency.
 - [Percentage][frequenz.quantities.Percentage]: A quantity representing a percentage.
 - [Power][frequenz.quantities.Power]: A quantity representing power.
+- [ReactivePower][frequenz.quantities.ReactivePower]: A quantity representing reactive
+  power.
 - [Temperature][frequenz.quantities.Temperature]: A quantity representing temperature.
 - [Voltage][frequenz.quantities.Voltage]: A quantity representing electric voltage.
-
-Additionally, for each of those types, there is a corresponding marshmallow
-field that can be used to serialize and deserialize the quantities using the
-[QuantitySchema][frequenz.quantities.marshmallow.QuantitySchema] schema.
 
 There is also the unitless [Quantity][frequenz.quantities.Quantity] class. All
 quantities are subclasses of this class and it can be used as a base to create new
@@ -78,6 +78,12 @@ Example:
     except TypeError as e:
         print(f"Error: {e}")  # Error: unsupported operand type(s) for +: 'Power' and 'Voltage'
     ```
+
+This library also provides an [**experimental** module with marshmallow fields and
+a base schema][frequenz.quantities.experimental.marshmallow] to serialize and
+deserialize quantities using the marshmallow library. To use it, you need to make sure
+to install this package with the `marshmallow` optional dependencies (e.g.
+`pip install frequenz-quantities[marshmallow]`).
 """
 
 
@@ -93,6 +99,7 @@ from ._temperature import Temperature
 from ._voltage import Voltage
 
 __all__ = [
+    "ApparentPower",
     "Current",
     "Energy",
     "Frequency",
@@ -100,7 +107,6 @@ __all__ = [
     "Power",
     "Quantity",
     "ReactivePower",
-    "ApparentPower",
     "Temperature",
     "Voltage",
 ]