+++ RELEASE_NOTES.md
@@ -9,5 +9,5 @@ This is the initial release, extracted from the [SDK v1.0.0rc601](https://github
 - Added support for `__round__` (`round(quantity)`), `__pos__` (`+quantity`) and `__mod__` (`quantity % quantity`) operators.
 - Add `ReactivePower` quantity.
 - Add `ApparentPower` quantity.
-- Add marshmallow module available when adding `[marshmallow]` to the requirements.
+- Add an **experimental** marshmallow module available when adding `[marshmallow]` to the requirements.
    * Add a QuantitySchema supporting string/float based serialization and deserialization of quantities.